split nse_update.H into Strang and Simplified-SDC versions this makes it more readable
